refactor: 공통 로직 병합
